% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhero_plot.R
\name{add_forecast_shading}
\alias{add_forecast_shading}
\title{Adds a gray shaded region to a ggplot chart to indicate forecasted data.}
\usage{
add_forecast_shading(plot, min_x, max_x)
}
\arguments{
\item{plot}{A ggplot object, such as one generated by `uhero_draw_dual_y_chart`}

\item{min_x}{The starting point along the x-axis.}

\item{max_x}{The ending point along the x-axis}
}
\value{
Returns a ggplot object with a rectangular annotation layer for
the forecasted data.
}
\description{
Adds a gray shaded region to a ggplot chart to indicate forecasted data.
}
\examples{
df <- data.frame(
  x = c(2010, 2011, 2012, 2013, 2014, 2015),
  y = c(5, 3, 10, 15, 4, 7),
  z = c(55, 70, 62, 58, 60, 68)
)
\dontrun{
chart <- uhero_draw_dual_y_chart(
  df,
  y1_series = c("y"),
  y2_series = c("z"),
  y1_limits = NULL,
  y2_limits = NULL,
  x_var = "x",
  y1_chart_type = "line",
  y2_chart_type = "line",
  y1_percent = FALSE,
  y2_percent = FALSE,
  y1_unit_prefix = '',
  y2_unit_prefix = ''
)
add_forecast_shading(chart, 2013, 2015)
}
}
