% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhero_plot.R
\name{add_text_labels}
\alias{add_text_labels}
\title{Replace ggplot legend with text labels}
\usage{
add_text_labels(plot, font_size = 3)
}
\arguments{
\item{plot}{A ggplot plot object}

\item{font_size}{A number for the font size of the labels. Defaults to 3.}
}
\value{
Returns a ggplot object with text labels.
}
\description{
Replace ggplot legend with text labels
}
\examples{
df <- data.frame(
  x = c(2010, 2011, 2012, 2013, 2014, 2015),
  y = c(5, 3, 10, 15, 4, 7),
  z = c(55, 70, 62, 58, 60, 68)
)
\dontrun{
example_plot <- uhero_draw_ggplot(
  data = df,
  series = c("y"),
  x_var = "x",
  y_limits = NULL,
  chart_type = "line",
  percent = FALSE,
  unit_prefix = '',
)
add_text_labels(exanple_plot)
}
}
