% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_charts.R
\name{draw_fcast_layout}
\alias{draw_fcast_layout}
\title{Preview chart for forecast layout}
\usage{
draw_fcast_layout(plot, w = 4.5, h = 2.45)
}
\arguments{
\item{plot}{Plot object}

\item{w}{Width of the view port in inches, defaults to 4.5}

\item{h}{Height of the view port in inches, defaults to 2.45}
}
\description{
This modifies the size of the view port window to more accurately reflect the size
of the figure including the placement of elements like data labels that would be used in a forecast layout.
}
\examples{
df <- data.frame(
  x = rnorm(10) * 100000,
  y = seq(0, 1, length.out = 10)
)
plot <- ggplot2::ggplot(df, ggplot2::aes(x, y)) + ggplot2::geom_point()
draw_fcast_layout(plot)
}
