% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhero_plot.R
\name{dual_y_axis_transform}
\alias{dual_y_axis_transform}
\title{Set up transformation and scaling functions for charts with two y axes.}
\usage{
dual_y_axis_transform(
  data,
  y1_series,
  y2_series,
  y1_limits = NULL,
  y2_limits = NULL
)
}
\arguments{
\item{data}{Dataset to be used. See sampleData.xlsx in the sampledata directory of the package for an example}

\item{y1_series}{A vector of string names (i.e. c("col_name")) corresponding to the columns in the data that should be plotted on the primary/left axis.}

\item{y2_series}{A vector of string names (i.e. c("col_name")) corresponding to the columns in the data that should be plotted on the secondary/right axis.}

\item{y1_limits}{A vector of numbers c(min, max, step) indicating the min, max, and steps between axis labels on the left axis. This is optional and defaults to NULL if not specified.}

\item{y2_limits}{A vector of numbers c(min, max, step) indicating the min, max, and steps between axis labels on the right axis. This is optional and defaults to NULL if not specified.}
}
\value{
Returns a list of two functions: rescale and transform.
Rescale is used to scale the data that is to be plotted on the secondary/right axis
to the data that is on the primary/left axis.
Transform is used to transform the secondary/right axis.
}
\description{
Set up transformation and scaling functions for charts with two y axes.
}
\examples{
df <- data.frame(
  x = c(2010, 2011, 2012, 2013, 2014, 2015),
  y = c(5, 3, 10, 15, 4, 7),
  z = c(55, 70, 62, 58, 60, 68)
)
dual_y_axis_transform(
  df,
  y1_series = c("y"),
  y2_series = c("z"),
  y1_limits = c(0, 15, 5),
  y2_limits = c(0, 70, 10)
)
}
