% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_charts.R
\name{export_report_layout}
\alias{export_report_layout}
\title{Export chart for UHERO report layout}
\usage{
export_report_layout(
  file_name,
  plot,
  w = 5.6931,
  h = 4,
  u = "in",
  bg = "transparent",
  ...
)
}
\arguments{
\item{file_name}{A string for the file name, including the extension.}

\item{plot}{Ggplot plot object.}

\item{w}{Integer - width of the exported image, defaults to 5.6931}

\item{h}{Integer - height of the exported image, defaults to 4}

\item{u}{A string for the units, defaults to "in" for inches}

\item{bg}{Background color. Defaults to "transparent". Use e.g. "white" for a
solid background, or NULL to keep the background set by the plot's theme.}

\item{...}{Additional parameters that can be passed to ggplot2::ggsave}
}
\description{
Uses \code{ggplot2::ggsave} to save a copy of the chart. By default, charts are sized at 5.6931 x 4 inches.
Sometimes charts may need to be larger, so the export function does accept parameters to change the width and height.
Please use a `.svg`, `.pdf`, or `.eps` extension when exporting for a report layout.
}
\examples{
df <- data.frame(
  x = rnorm(10) * 100000,
  y = seq(0, 1, length.out = 10)
)
plot <- ggplot2::ggplot(df, ggplot2::aes(x, y)) + ggplot2::geom_point()
export_report_layout('plot.svg', plot)
}
