% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhero_colors.R
\name{uhero_colors}
\alias{uhero_colors}
\title{Hex codes for a given color in the UHERO colors}
\usage{
uhero_colors(...)
}
\arguments{
\item{...}{Names of the UHERO colors. There are 14 total:
"blue", "orange", "green", "purple", "cyan", "gray", "red", "light blue", "light orange", "light green", "light purple",
"light cyan", "light gray", and "light red".}
}
\value{
A list of colors with their HEX codes.
}
\description{
Hex codes for a given color in the UHERO colors
}
\examples{
uhero_colors("blue")
uhero_colors("blue", "light orange")
}
