% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhero_plot.R
\name{uhero_draw_dual_y_chart}
\alias{uhero_draw_dual_y_chart}
\title{Generate a plot with ggplot2 with the UHERO theme with 2 y axes.}
\usage{
uhero_draw_dual_y_chart(
  data,
  y1_series,
  y2_series,
  y1_limits = NULL,
  y2_limits = NULL,
  x_var,
  y1_chart_type = "line",
  y2_chart_type = "line",
  y1_percent = FALSE,
  y2_percent = FALSE,
  y1_unit_prefix = "",
  y2_unit_prefix = ""
)
}
\arguments{
\item{data}{Dataset to be used. See sampleData.xlsx in the sampledata directory of the package for an example.}

\item{y1_series}{A vector of string names (i.e. c("col_name")) corresponding to the columns in the data that should be plotted on the primary/left axis.}

\item{y2_series}{A vector of string names (i.e. c("col_name")) corresponding to the columns in the data that should be plotted on the secondary/right axis.}

\item{y1_limits}{A vector of numbers c(min, max, step) indicating the min, max, and steps between axis labels on the left axis. This is optional and defaults to NULL if not specified.}

\item{y2_limits}{A vector of numbers c(min, max, step) indicating the min, max, and steps between axis labels on the right axis. This is optional and defaults to NULL if not specified.}

\item{x_var}{A string corresponding to the column that is to be used for the x axis}

\item{y1_chart_type}{A string, either "line" or "col" indicating the chart type to be used for series on the left axis.}

\item{y2_chart_type}{A string, either "line" or "col" indicating the chart type to be used for series on the right axis.}

\item{y1_percent}{A boolean indicating if the maximum left axis label should be marked with a "\%" sign. Defaults to FALSE.}

\item{y2_percent}{A boolean indicating if the maximum right axis label should be marked with a "\%" sign. Defaults to FALSE.}

\item{y1_unit_prefix}{A string to define any symbols that should be in front of the maximum left axis label (i.e. "$"). Defaults to an empty string.}

\item{y2_unit_prefix}{A string to define any symbols that should be in front of the maximum right axis label (i.e. "$"). Defaults to an empty string.}
}
\value{
Returns a ggplot object
}
\description{
Generate a plot with ggplot2 with the UHERO theme with 2 y axes.
}
\examples{
df <- data.frame(
  x = c(2010, 2011, 2012, 2013, 2014, 2015),
  y = c(5, 3, 10, 15, 4, 7),
  z = c(55, 70, 62, 58, 60, 68)
)
\dontrun{uhero_draw_dual_y_chart(
  df,
  y1_series = c("y"),
  y2_series = c("z"),
  y1_limits = NULL,
  y2_limits = NULL,
  x_var = "x",
  y1_chart_type = "line",
  y2_chart_type = "line",
  y1_percent = FALSE,
  y2_percent = FALSE,
  y1_unit_prefix = '',
  y2_unit_prefix = ''
)}
}
