% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhero_colors.R
\name{uhero_scale_colour_diverge}
\alias{uhero_scale_colour_diverge}
\title{Applies the UHERO color palette to ggplot's diverging color aesthetic.}
\usage{
uhero_scale_colour_diverge(high = "blue", low = "orange", ...)
}
\arguments{
\item{high}{A string that defaults to "blue". Can accept any of the other colors in \code{uhero_color_list}.}

\item{low}{A string that defaults to "orange". Can accept any of the other colors in \code{uhero_color_list}.}

\item{...}{Any other parameters that can be passed to \code{ggplot2::scale_color_gradient2}}
}
\description{
Uses \code{ggplot2::scale_color_gradient2}. \code{uhero_scale_color_diverge} is available as an alias.
}
\examples{
set.seed(1)
df <- data.frame(
 x = runif(100),
 y = runif(100),
 z1 = rnorm(100),
 z2 = abs(rnorm(100))
)
ggplot2::ggplot(df, ggplot2::aes(x, y)) +
ggplot2::geom_point(ggplot2::aes(colour = z1)) +
uhero_scale_colour_diverge()
}
