% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhero_colors.R
\name{uhero_scale_fill}
\alias{uhero_scale_fill}
\title{Applies the UHERO color palette to ggplot's fill aesthetic.}
\usage{
uhero_scale_fill(palette = "primary", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{A string that defaults to "primary". This can also be set to either "secondary" or "all".}

\item{discrete}{A boolean that defaults to TRUE. Set to FALSE for a continuous scale.}

\item{reverse}{A boolean that defaults to FALSE. Set to TRUE to reverse the color scale.}

\item{...}{Additional parameters that can be passed to \code{ggplot2::scale_fill_gradientn}}
}
\description{
Uses \code{ggplot2::discrete_scale} for the fill aesthetic for discrete scales and \code{ggplot2::scale_fill_gradientn}
for the fill aesthetic on continuous scales.
}
\examples{
ggplot2::ggplot(ggplot2::mpg, ggplot2::aes(hwy, fill = class)) +
ggplot2::geom_bar(show.legend = FALSE) +
uhero_scale_fill(palette = 'secondary')
}
