% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhero_colors.R
\name{uhero_scale_fill_diverge}
\alias{uhero_scale_fill_diverge}
\title{Applies the UHERO color palette to ggplot's diverging fill aesthetic.}
\usage{
uhero_scale_fill_diverge(high = "blue", low = "orange", ...)
}
\arguments{
\item{high}{A string that defaults to "blue". Can accept any of the other colors in \code{uhero_color_list}.}

\item{low}{A string that defaults to "orange". Can accept any of the other colors in \code{uhero_color_list}.}

\item{...}{Any other parameters that can be passed to \code{ggplot2::scale_fill_gradient2}}
}
\description{
Uses \code{ggplot2::scale_fill_gradient2}
}
