% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UHEROtheme-package.R
\docType{package}
\name{UHEROtheme-package}
\alias{UHEROtheme}
\alias{UHEROtheme-package}
\title{UHEROtheme: UHERO Theme For Ggplot Charts}
\description{
Creates a theme and utility functions to help export charts following the UHERO style guide.
}
\author{
\strong{Maintainer}: Victoria Rhinebolt \email{vward@hawaii.edu}

}
\keyword{internal}
