% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_charts.R
\name{export_plot}
\alias{export_plot}
\title{Export plot}
\usage{
export_plot(file_name, plot, w = 1920, h = 1080, u = "px", ...)
}
\arguments{
\item{file_name}{A string for the file name, including the extension.}

\item{plot}{Ggplot plot object.}

\item{w}{Integer - width of the exported image, defaults to 1920}

\item{h}{Integer - height of the exported image, defaults to 1080}

\item{u}{A string for the units, defaults to "px" for pixels}

\item{...}{Additional parameters that can be passed to ggplot2::ggsave}
}
\description{
Uses \code{ggplot2::ggsave} to save a copy of the chart. By default the exports are sized at 1920 x 1080 pixels.
}
\examples{
df <- data.frame(
  x = rnorm(10) * 100000,
  y = seq(0, 1, length.out = 10)
)
plot <- ggplot2::ggplot(df, ggplot2::aes(x, y)) + ggplot2::geom_point()
export_report_layout('plot.png', plot)
}
