% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhero_plot.R
\name{uhero_draw_dual_y_ggplot}
\alias{uhero_draw_dual_y_ggplot}
\title{Generate a plot with ggplot2 with the UHERO theme with 2 y axes.}
\usage{
uhero_draw_dual_y_ggplot(
  data,
  x_var,
  y1 = list(series = NULL, chart_type = "line", limits = NULL, percent = FALSE,
    unit_prefix = ""),
  y2 = list(series = NULL, chart_type = "line", limits = NULL, percent = FALSE,
    unit_prefix = ""),
  ...
)
}
\arguments{
\item{data}{Dataset to be used. See sampleData.xlsx in the sampledata directory of the package for an example.}

\item{x_var}{A string corresponding to the column that is to be used for the x axis}

\item{y1}{A list specifying the series to be plotted on the primary/left axis. Items in the list are:
`series` A vector of string names (i.e. c("col_name")) corresponding to the columns in the data. Defaults to NULL
`chart_type` A string, either "line" or "col", or a vector
(i.e., c("col_name1" = line", "col_name2" = "bar")) to indicate how the series should be drawn.
Defaults to "line".
`limits` A vector of numbers c(min, max, step) indicating the min, max, and steps between axis
labels. This is optional and defaults to NULL if not specified.
`percent` A boolean indicating if the axis is in percentages. Defaults to FALSE.
`unit_prefix` A string to define any symbols that should be in front of the
maximum axis label (i.e. "$"). Defaults to an empty string.}

\item{y2}{A list specifying the series to be plotted on the secondary/right axis. Items in the list
and their defaults are the same as the `y1` parameter}

\item{...}{Additional optional parameters that can be used by ggplot geoms. For example, `position = "dodge2"` for a bar chart.}
}
\value{
A structure with the plot and scaled data used in the plot
}
\description{
Generate a plot with ggplot2 with the UHERO theme with 2 y axes.
}
\examples{
df <- data.frame(
  x = c(2010, 2011, 2012, 2013, 2014, 2015),
  y = c(5, 3, 10, 15, 4, 7),
  z = c(55, 70, 62, 58, 60, 68)
)
\dontrun{uhero_draw_dual_y_ggplot(
  df,
  x_var = "x",
  y1 = list(
    series = c("y")
  ),
  y2 = list(
    series = c("z")
  )
)}
}
