% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhero_plot.R
\name{uhero_draw_ggplot}
\alias{uhero_draw_ggplot}
\title{Generates a ggplot2 chart with the UHERO theme.}
\usage{
uhero_draw_ggplot(
  data,
  series,
  x_var,
  y_limits = NULL,
  chart_type = NULL,
  percent = FALSE,
  unit_prefix = "",
  ...
)
}
\arguments{
\item{data}{Dataset to be used. See sampleData.xlsx in the sampledata directory of the package for an example.}

\item{series}{A vector of string names (i.e. c("col_name")) corresponding to the columns in the data that should be plotted on the primary/left axis.}

\item{x_var}{A string corresponding to the column that is to be used for the x axis}

\item{y_limits}{A vector of numbers c(min, max, step) indicating the min, max, and steps between axis labels on the left axis. This is optional and defaults to NULL if not specified.}

\item{chart_type}{A string or a vector indicating the chart type for each series. A single string like "bar" can be used to apply the bar chart type to all series.
A vector such as c("line", "bar") or c("Col Name" = "line", "Col Name" = "bar") can be used to specify a type for each series. Current accepted chart type values
are "line", "bar", and "col". Defaults to NULL. If not specified the plot will default to a line chart.}

\item{percent}{A boolean indicating if the maximum y axis label should be marked with a "\%" sign. Defaults to FALSE.}

\item{unit_prefix}{A string to define any symbols that should be in front of the maximum y axis label (i.e. "$"). Defaults to an empty string.}

\item{...}{Additional optional parameters that can be used by ggplot geoms. For example, `position = "dodge2"` for a bar chart.}
}
\value{
A structure with the plot and scaled data used in the plot
}
\description{
Generates a ggplot2 chart with the UHERO theme.
}
\examples{
df <- data.frame(
  x = c(2010, 2011, 2012, 2013, 2014, 2015),
  y = c(5, 3, 10, 15, 4, 7),
  z = c(55, 70, 62, 58, 60, 68)
)
\dontrun{
uhero_draw_ggplot(
  data = df,
  series = c("y"),
  x_var = "x",
  y_limits = NULL,
  chart_type = "line",
  percent = FALSE,
  unit_prefix = '',
)
}
}
