% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhero_plot.R
\name{uhero_draw_ggplot}
\alias{uhero_draw_ggplot}
\title{Generates a ggplot2 chart with the UHERO theme.}
\usage{
uhero_draw_ggplot(
  data,
  series,
  x_var,
  y_limits = NULL,
  chart_type = "line",
  bar_position = NULL,
  percent = FALSE,
  unit_prefix = ""
)
}
\arguments{
\item{data}{Dataset to be used. See sampleData.xlsx in the sampledata directory of the package for an example.}

\item{series}{A vector of string names (i.e. c("col_name")) corresponding to the columns in the data that should be plotted on the primary/left axis.}

\item{x_var}{A string corresponding to the column that is to be used for the x axis}

\item{y_limits}{A vector of numbers c(min, max, step) indicating the min, max, and steps between axis labels on the left axis. This is optional and defaults to NULL if not specified.}

\item{chart_type}{A string, either "line", "col", or "bar" indicating the chart type to be used for the series.}

\item{bar_position}{A string to indicate the position of the bars ("stack", "fill", "dodge", "dodge2")
if the chart type specified is set to "bar". Defaults to NULL resulting in a stacked bar chart.}

\item{percent}{A boolean indicating if the maximum y axis label should be marked with a "\%" sign. Defaults to FALSE.}

\item{unit_prefix}{A string to define any symbols that should be in front of the maximum y axis label (i.e. "$"). Defaults to an empty string.}
}
\value{
Returns a ggplot object
}
\description{
Generates a ggplot2 chart with the UHERO theme.
}
\examples{
df <- data.frame(
  x = c(2010, 2011, 2012, 2013, 2014, 2015),
  y = c(5, 3, 10, 15, 4, 7),
  z = c(55, 70, 62, 58, 60, 68)
)
\dontrun{uhero_draw_ggplot(
  data = df,
  series = c("y"),
  x_var = "x",
  y_limits = NULL,
  chart_type = "line",
  bar_position = NULL,
  percent = FALSE,
  unit_prefix = '',
)}
}
