% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhero_theme.R
\name{uhero_theme}
\alias{uhero_theme}
\title{Adds a UHERO theme for Ggplot charts.}
\usage{
uhero_theme(layout = FALSE)
}
\arguments{
\item{layout}{A boolean to indicate whether or not the theme is being applied to a plot that is to be used in a UHERO report layout.
This defaults to false. If layout is TRUE, then the base size is set to 9 and the font family is set to open sans. If the parameter is FALSE, then
the base size is set to 30 and the font family is set to opensans-semibold. Set layout to TRUE if the plot is being used for a UHERO report.}
}
\description{
Modifies the ggplot minimal theme to fit the style used for UHERO reports/presentations.
}
\examples{
plot <- ggplot2::ggplot(ggplot2::mpg) + uhero_theme()
}
