% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhero_theme.R
\name{uhero_theme}
\alias{uhero_theme}
\title{Adds a UHERO theme for Ggplot charts.}
\usage{
uhero_theme(font_size = 9)
}
\arguments{
\item{font_size}{A number indicating the font size to be used by ggplot's element_text. Defaults to 9.}
}
\description{
Modifies the ggplot minimal theme to fit the style used for UHERO reports/presentations.
}
\examples{
plot <- ggplot2::ggplot(ggplot2::mpg) + uhero_theme()
}
